src/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aws_lambda_handler_cookbook"
version = "3.0.0"
description = "This repository provides a working, deployable, open source based, AWS Lambda handler and CDK Python code. This handler embodies Serverless best practices and has all the bells and whistles for a proper production ready handler"
authors = [{name = "Ran Isenberg"}]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.14",
]
readme = "README.md"
keywords = ["aws lambda cookbook", "guide to aws lambda", "cdk blueprint", "best practices", "serverless service blueprint"]
license = {text = "MIT-0"}
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.12.0",
    "email-validator",
    "aws-lambda-powertools[tracer]>=3.22.1",
    "mypy-boto3-dynamodb",
    "cachetools",
    "boto3>=1.26.125",
    "aws-lambda-env-modeler",
    # CareerVP specific
    "anthropic>=0.40.0",
    "beautifulsoup4>=4.12.0",
    "langdetect>=1.0.9",
    "python-docx>=1.1.0",
    "PyPDF2>=3.0.0",
    "httpx>=0.27.0",
    "moto>=5.1.20",
    "pytest-asyncio>=1.3.0",
]

[project.urls]
Repository = "https://github.com/ran-isenberg/aws-lambda-handler-cookbook"

[dependency-groups]
dev = [
    # CDK
    "service-cdk",
    "aws-cdk-lib>=2.180.0",
    "constructs>=10.0.0",
    "cdk-nag>2.0.0",
    "aws-cdk-aws-lambda-python-alpha>=2.180.0a0",
    "cdk-monitoring-constructs",
    # DEV
    "pytest",
    "pytest-mock",
    "moto[dynamodb,s3]>=5.0.0",
    "pycodestyle",
    "pytest-cov",
    "pytest-html",
    "python-dateutil",
    "python-dotenv",
    "GitPython",
    "ruff",
    "radon",
    "xenon",
    "pre-commit",
    "mkdocs-material",
    "mkdocs-git-revision-date-plugin",
    "mkdocs-glightbox",
    "types-cachetools",
    "mypy",
    "types-requests",
    "toml",
    "mkdocs-render-swagger-plugin",
]

[tool.uv.sources]
service-cdk = { path = "../../infra", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["service"]

[tool.hatch.build]
skip-excluded-dirs = true

[tool.pytest.ini_options]
testpaths = "tests"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "cdk.out",
    ".vscode",
    ".pytest_cache",
    ".build",
    ".mypy_cache",
    ".github"
]

# Same as Black.
line-length = 150
indent-width = 4

# Assume Python 3.14
target-version = "py314"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [ "E203", "E266", "E501", "W191"]

[tool.ruff.format]
quote-style = "single"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.isort]
known-third-party = ["pydantic", "aws_lambda_powertools"]
